DMF : refactored generalizable routines to LTDL the LibNEON and initiate parsing into dmfcore.c/.h separately from specific dmfsnmp.c/.h code; fixed a bug in dmf_parse_dir() vs absent directory
