+++ include/dmf.h
@@ -27,5 +27,11 @@
  * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA
  */
 
+/* Note: code that includes this must be built with $(LIBNEON_CFLAGS) */
+#ifndef DMF_H
+#define DMF_H
+
+#include "dmfcore.h"
 #include "dmfsnmp.h"
 
+#endif /* DMF_H */
